take out filters from commons.js (wip)
