+++ app/javascript/gobierto_investments/webapp/containers/home/Home.vue
@@ -115,8 +115,13 @@ import RangeBars from "../../components/RangeBars.vue";
 import axios from "axios";
 
 import { BlockHeader, Calendar } from "lib/vue-components";
+import Middleware from "lib/middleware";
+
 import { CommonsMixin, baseUrl } from "../../mixins/common.js";
 import { store } from "../../mixins/store";
+
+// TODO: This configuration should come from API request, not from file
+import CONFIGURATION from "./mataro.conf.js";
 
 export default {
   name: "Home",
@@ -190,8 +195,6 @@ export default {
         axios.get(`${baseUrl}/meta?stats=true`)
       ]);
 
-      this.dictionary = attributesDictionary;
-
       let items = this.setData(__items__);
       let phases = [];
       let filters = [];
@@ -212,7 +215,8 @@ export default {
         // Add dictionary of phases in order to fulfill project page
         items = items.map(item => ({ ...item, phasesDictionary: __phases__ }));
 
-        filters = this.getFilters(filtersFromConfiguration) || [];
+        const middleware = new Middleware({ dictionary: attributesDictionary, availableFilters: CONFIGURATION })
+        filters = middleware.getFilters(filtersFromConfiguration) || [];
 
         if (filters.length) {
           this.activeFilters = new Map();
