+++ app/javascript/lib/shared/modules/vue-filters.js
@@ -9,26 +9,32 @@
  * mixins: [VueFiltersMixin],
  *
  */
-const translate = (value = {}) => {
+export const translate = (value = {}) => {
   const lang = I18n.locale || "es";
   // Look for the language key, fallback to the first found key
   return (Object.prototype.hasOwnProperty.call(value, lang) && value[lang]) ? value[lang] : value[Object.keys(value)[0]];
 }
 
+export const money = (value, opts = {}) => {
+  const lang = I18n.locale || "es";
+  const options = { style: "currency", currency: "EUR", ...opts }
+  return value !== undefined && value !== null ? parseFloat(value).toLocaleString(lang, options) : undefined;
+}
+
+export const date = (value, opts = {}) => {
+  const lang = I18n.locale || "es";
+  return value instanceof Date ? value.toLocaleDateString(lang, opts) : new Date(value).toLocaleDateString(lang, opts);
+}
+
 export const VueFiltersMixin = {
   methods: {
-    translate
+    translate,
+    money,
+    date
   },
   filters: {
     translate,
-    money(value, opts = {}) {
-      const lang = I18n.locale || "es";
-      const options = { style: "currency", currency: "EUR", ...opts }
-      return value !== undefined && value !== null ? parseFloat(value).toLocaleString(lang, options) : undefined;
-    },
-    date(value, opts = {}) {
-      const lang = I18n.locale || "es";
-      return value instanceof Date ? value.toLocaleDateString(lang, opts) : new Date(value).toLocaleDateString(lang, opts);
-    }
+    money,
+    date
   }
 };
